DoReMi_subscription/requirements.txt
DateTime==4.4
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
zope.interface==5.4.0
